fixed compile errors and warnings
<==== Mercurial data ====>
Branch: x_CLOUD-536_verify_transaction_order

+++ common/src/utils/db/types.h
@@ -47,7 +47,7 @@ public:
     QString password;
     QString connectOptions;
     QString encoding;
-    size_t maxConnectionCount;
+    int maxConnectionCount;
     /** Connection is closed if not used for this interval. */
     std::chrono::seconds inactivityTimeout;
     /**
